download.sh: add "--no-arbitrary-wait" option, you can tell plowdown to give up you received a "come back later" message (you don't know how much time to wait)
rapidshare & megaupload modules now use that.

See also issue 239

git-svn-id: https://plowshare.googlecode.com/svn/trunk@1256 51fab9ec-ea3e-11dd-80ed-bda8ba2650cd

+++ src/modules/megaupload.sh
@@ -122,9 +122,12 @@ megaupload_download() {
 
         # Test for "come back later". Language is guessed with the help of http-user-agent.
         elif match 'file you are trying to access is temporarily unavailable' "$PAGE"; then
-            log_debug "File temporarily unavailable"
-            rm -f $COOKIES
-            return 253
+            if test "$NOARBITRARYWAIT"; then
+                log_debug "File temporarily unavailable"
+                rm -f $COOKIES
+                return 253
+            fi
+            continue
         fi
 
         # Look for a download link (usually a password protected file)
